Wind speed decimal place OK

+++ web/src/features/fireBehaviourCalculator/tableState.ts
@@ -48,7 +48,7 @@ export const buildUpdatedNumberRow = (
   return {
     ...rowToUpdate,
     ...{
-      [field as keyof FBCInputRow]: parseInt(value)
+      [field as keyof FBCInputRow]: parseFloat(value)
     }
   }
 }
